requirements.txt
flask
langchain
langchain-openai
wikipedia
python-pptx
python-dotenv
waitress



